Completely change the parallelization model.
It turns out that all OpenGL stuff using the same context must be in the same
thread. Now, there is almost no parallelization, except that the main thread
can do other things while all images are generated and copied to memory.
It is possible to extend that by adding more copy threads, and waiting for
them to finish before setting the promise.

I don't implement it now because the whole render thread / tiled view
interaction should be rethought. No reason to create a context on each
new render operation.